requirements.txt
telebot
coloredlogs
pyTelegramBotAPI
python-dotenv
pymongo
sshtunnel